made workflow.tasks a submodule
